Rename: recoil 폴더 구조 변경
- recoil
	- atoms
		- auth.ts
		- darkmode.ts. // 예시
	- selectors  // 현재는 없음
