+++ src/pages/HomePage.tsx
@@ -7,7 +7,7 @@ import { unOfficialGroupDummyData } from '@dummy/group';
 import { ReactComponent as Logo } from '@assets/images/logo/logo.svg';
 import SearchInput from '@components/SearchInput';
 import { useRecoilValue } from 'recoil';
-import isLoggedInState from '@recoil/login/atoms';
+import isLoggedInState from '@recoil/atoms/auth';
 import { useNavigate } from 'react-router-dom';
 
 const titleStyle = 'font-bold text-lg mb-4 mt-20';
