--- src/recoil/login/atoms.ts
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-const isLoggedInState = atom<boolean>({
-  key: 'isLoggedInState',
-  default: true,
-});
-
-export default isLoggedInState;
